+++ pages/boards/index.tsx
@@ -1,11 +1,10 @@
 import ListContainer from "../../src/component/unit/boards/list/list.container";
-import ListContainer2 from "../../src/component/unit/boards/list/list.container copy";
 
 // list
 export default function Boards() {
   return (
     <>
-      <ListContainer2 />
+      <ListContainer />
     </>
   );
 }
