+++ src/component/unit/boards/detail/detail.container.tsx
@@ -84,6 +84,31 @@ export default function DetailContainer() {
   };
 
   // 좋아요 함수
+  // const onClickLike = async (event: MouseEvent<HTMLImageElement>) => {
+  //   try {
+  //     const result = await clicklike({
+  //       variables: {
+  //         boardId: String((event.target as HTMLImageElement).id),
+  //       },
+  //       update(cache, { data }) {
+  //         const fetchUserLike = data?.fetchUserLike;
+  //         cache.modify({
+  //           fields: {
+  //             fetchUserLike: (prev, { readField }) => {
+  //               const filteredPrev = prev.filter(
+  //                 (el: any) => readField("id", el) !== fetchUserLike
+  //               );
+  //               return [...filteredPrev];
+  //             },
+  //           },
+  //         });
+  //       },
+  //     });
+  //     setIsActive(result.data?.clickLike);
+  //   } catch (error) {
+  //     alert("실패");
+  //   }
+  // };
   const onClickLike = async (event: MouseEvent<HTMLImageElement>) => {
     try {
       const result = await clicklike({
